simply tack on another dialog state Fixes #4994
The alternate implementation here is to check if the current pages
data-url is already a part of the hash and replace it with just
the dialog hash key. This is dangerous though because it doesn't
account for dialog states either. This solution is simple and works
with the only downside being an odd url in an extreme corner case.